+++ app/src/main/java/com/pzbdownloaders/scribble/about_us/presentation/screens/AboutUsScreen.kt
@@ -1,8 +1,31 @@
 package com.pzbdownloaders.scribble.about_us.presentation.screens
 
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.sp
+import com.pzbdownloaders.scribble.R
 
 @Composable
-fun AboutUs() {
-
+fun AboutUsScreen() {
+    Column(
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally,
+        modifier = Modifier.fillMaxSize()
+    ) {
+        Text(text = "ABOUT US", color = MaterialTheme.colors.onPrimary, fontSize = 25.sp)
+        Icon(
+            painter = painterResource(id = R.drawable.ic_launcher),
+            contentDescription = "icon",
+        )
+    }
 }