+++ app/src/main/java/com/pzbdownloaders/scribble/archive_notes_feature/presentation/screen/ArchiveNotesScreen.kt
@@ -1,6 +1,8 @@
 package com.pzbdownloaders.scribble.archive_notes_feature.presentation.screen
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.livedata.observeAsState
 import androidx.navigation.NavHost
 import androidx.navigation.NavHostController
 import com.pzbdownloaders.scribble.archive_notes_feature.presentation.components.MainStructureArchiveScreen
@@ -13,10 +15,21 @@ fun ArchiveNotesScreen(
     navHostController: NavHostController,
     viewModel: MainActivityViewModel,
     activity: MainActivity,
+    selectedItem: MutableState<Int>,
+    selectedNote: MutableState<Int>
 ) {
+    viewModel.getNoteBook()
+    val notebooks = viewModel.getNoteBooks.observeAsState().value
+    val notebookNavigation: ArrayList<String> = ArrayList()
+    for (i in notebooks?.indices ?: emptyList<String>().indices) {
+        notebookNavigation.add(notebooks!![i]?.notebook ?: "")
+    }
     MainStructureArchiveScreen(
         navHostController = navHostController,
         viewModel = viewModel,
-        activity = activity
+        activity = activity,
+        notebookNavigation,
+        selectedItem,
+        selectedNote
     )
 }