+++ app/src/main/java/com/pzbdownloaders/scribble/di/AppModule.kt
@@ -16,6 +16,7 @@ import com.pzbdownloaders.scribble.login_and_signup_feature.domain.usecase.SignI
 import com.pzbdownloaders.scribble.login_and_signup_feature.domain.usecase.SignUpUserCase
 import com.pzbdownloaders.scribble.main_screen.data.repository.NoteRepository
 import com.pzbdownloaders.scribble.main_screen.domain.usecase.GetNotesUseCase
+import com.pzbdownloaders.scribble.notebook_main_screen.domain.GetNotebookNotesUseCase
 import com.pzbdownloaders.scribble.search_feature.domain.usecase.GetArchiveSearchResultUseCase
 import com.pzbdownloaders.scribble.search_feature.domain.usecase.GetSearchResultUseCase
 import dagger.Module
@@ -96,6 +97,9 @@ class AppModule {
 
     @Provides
     fun getNoteBookUseCase() = GetNoteBookUseCase()
+
+    @Provides
+    fun getNotebookNotesUseCase() = GetNotebookNotesUseCase()
 
 
 }