+++ app/src/main/java/com/pzbdownloaders/scribble/notebook_main_screen/presentation/screen/NotebookMainScreen.kt
@@ -1,5 +1,4 @@
-package com.pzbdownloaders.scribble.main_screen.presentation.screens
-
+package com.pzbdownloaders.scribble.notebook_main_screen.presentation.screen
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
@@ -8,17 +7,17 @@ import androidx.navigation.NavHostController
 import com.pzbdownloaders.scribble.common.presentation.MainActivity
 import com.pzbdownloaders.scribble.common.presentation.MainActivityViewModel
 import com.pzbdownloaders.scribble.main_screen.presentation.components.MainStructureMainScreen
+import com.pzbdownloaders.scribble.notebook_main_screen.presentation.components.MainStructureNotebookScreen
 
 @Composable
-fun NotesScreen(
+fun NotebookMainScreen(
     navHostController: NavHostController,
     viewModel: MainActivityViewModel,
     activity: MainActivity,
+    title: String,
     selectedItem: MutableState<Int>,
     selectedNote: MutableState<Int>
 ) {
-
-    // WE GET THE NOTEBOOKS HERE BECAUSE WE HAVE TO DISPLAY THEM IN THE NAVIGATION  VIEW
 
     viewModel.getNoteBook()
     val notebooks = viewModel.getNoteBooks.observeAsState().value
@@ -26,13 +25,13 @@ fun NotesScreen(
     for (i in notebooks?.indices ?: emptyList<String>().indices) {
         notebookNavigation.add(notebooks!![i]?.notebook ?: "")
     }
-    MainStructureMainScreen(
+    MainStructureNotebookScreen(
         navHostController,
         viewModel,
         activity,
         notebookNavigation,
+        title,
         selectedItem,
         selectedNote
     )
 }
-